+++ src/components/SharedLayout/SharedLayout.jsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
-import { Header } from './Header';
-import { SideBar } from './SideBar';
+import { Header } from '../Header/Header';
+import { SideBar } from '../SideBar/SideBar';
 
 export const SharedLayout = () => {
   return (
