--- src/pages/AccountPage.jsx
@@ -1,3 +1,0 @@
-export const AccountPage = () => {
-  return <div>account page</div>;
-};
