--- src/pages/Calendar.jsx
@@ -1,7 +1,0 @@
-export const Calendar = () => {
-  return (
-    <>
-      <div>calendar</div>
-    </>
-  );
-};
