--- src/pages/LoginPage.jsx
@@ -1,10 +1,0 @@
-import { Link } from 'react-router-dom';
-
-export const LoginPage = () => {
-  return (
-    <>
-      <div>login page</div>
-      <Link to="/register">Sign up</Link>
-    </>
-  );
-};
