--- src/pages/RegistrationPage.jsx
@@ -1,10 +1,0 @@
-import { Link } from 'react-router-dom';
-
-export const RegistrationPage = () => {
-  return (
-    <>
-      <div>registration page</div>
-      <Link to="/login">Log in</Link>
-    </>
-  );
-};
