--- src/pages/SignupPage.jsx
@@ -1,3 +1,0 @@
-export const SignupPage = () => {
-  return <div>signup page</div>;
-};
