--- src/pages/Statistics.jsx
@@ -1,7 +1,0 @@
-export const Statistics = () => {
-  return (
-    <>
-      <div>statistics</div>
-    </>
-  );
-};
